backend/requirements.txt
fastapi
uvicorn[standard]
pymongo
passlib[bcrypt]
python-jose
python-dotenv
python-multipart